requirements.txt
fastapi
uvicorn
python-multipart
pydantic
openai
PyMuPDF
python-docx
python-pptx
pytesseract
Pillow
python-dotenv
